Revert "spike(parity-judge): generalize semantic-equivalence judge to spike-D (tier-B parity) (#804)"
This reverts commit 0c71fbe3ab7fab2595fac797ebd72f7aa6932cd5.
